+++ src/corridor_planner.hh
@@ -77,7 +77,7 @@ namespace corridor_planner
         // If this value is true, the StrongEdgeAnnotation filter will be
         // passed on the resulting corridors
         bool strong_edge_enable;
-        envire::Grid<double>::Ptr strong_edge_map;
+        envire::Grid<float>::Ptr strong_edge_map;
         std::string strong_edge_band;
         double strong_edge_threshold;
 
@@ -126,7 +126,7 @@ namespace corridor_planner
         /** Annotates the resulting corridors using the StrongEdgeAnnotation
          * filter
          */
-        void enableStrongEdgeFilter(envire::Grid<double>::Ptr step_size, std::string const& band_name, double threshold);
+        void enableStrongEdgeFilter(envire::Grid<float>::Ptr step_size, std::string const& band_name, double threshold);
 
         /** Disables the strong edge filter as enabled by enableStrongEdgeFilter
          */
